Customize auto-configuration with properties
This commit showcases the use of `@ConfigurationProperties` as a means to
create a namespace for keys that are relevant to our auto-configuration.

We expose two new keys, `hello.prefix` and `hello.suffix`. We can then
easily use them in our auto-configuration and customize the prefix in our
app.

Unfortunately, the IDE is complaining that the key is unknown. Let's fix
that in the next commit!